+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Focus the created tab after the render.
+- Variables files to avoid prompt for some variables.
 
 ## [1.0.0] - 2022-09-19
 ### Added